Remove container images even if they are corrupted
It turns out that docker inspect fails to get the image id in cases
where the images are corrupted:
  # docker inspect --format='{{.Id}}' "${SUPERVISOR_IMAGE}"

  Error response from daemon: readlink /mnt/data/docker/overlay2: invalid argument

Fix this by using docker image instead of docker inspect which seems to
work even in cases where the image is currupted.

Also make sure to create the marker file before creating the container
or fetching the container image since it might be that container creation
or container image fetching actually fails due to a corrupted image.