Use performance governor by default
The performance governor seems not to have a major impact on power
consumption on any of the ARM platforms. Enable it by default and
remove the ODROID-N2+ special casing.

Leave Tinkerer ondemand as it seems to make almost 0.8W difference
between performance and ondemand, and along with that 8°C temperature
difference!